web-app/requirements.txt
flask
flask_wtf
email_validator
python-dotenv